+++ object/alive/animal/behavior/ai/memory.go
@@ -5,12 +5,12 @@ import "agar-life/math/crd"
 type memory struct {
 	valid     bool
 	priority  uint8
-	validTime uint64
+	validTime uint
 	reason    string
 	crd       crd.Crd
 }
 
-func (m *memory) set(pr uint8, vt uint64, reason string, crd crd.Crd) {
+func (m *memory) set(pr uint8, vt uint, reason string, crd crd.Crd) {
 	m.valid = true
 	m.priority = pr
 	m.validTime = vt
@@ -18,7 +18,7 @@ func (m *memory) set(pr uint8, vt uint64, reason string, crd crd.Crd) {
 	m.crd = crd
 }
 
-func (m *memory) check(pr uint8, cycle uint64) (bool, crd.Crd) {
+func (m *memory) check(pr uint8, cycle uint) (bool, crd.Crd) {
 	if m.valid && m.validTime < cycle && m.priority >= pr {
 		return true, m.crd
 	}
@@ -26,7 +26,7 @@ func (m *memory) check(pr uint8, cycle uint64) (bool, crd.Crd) {
 	return false, m.crd
 }
 
-func (m *memory) checkByReason(pr uint8, cycle uint64, reason string) (bool, crd.Crd) {
+func (m *memory) checkByReason(pr uint8, cycle uint, reason string) (bool, crd.Crd) {
 	if m.valid && m.validTime > cycle && m.priority >= pr && m.reason == reason {
 		return true, m.crd
 	}
