+++ object/alive/animal/behavior/follower.go
@@ -19,7 +19,7 @@ func NewFollower() animal.Behavior {
 	return &follower{Name: FollowerName}
 }
 
-func (a *follower) Action(self animal.Animal, animals []alive.Alive, plants []alive.Alive, cycle uint64) (crd.Crd, bool) {
+func (a *follower) Action(self animal.Animal, animals []alive.Alive, plants []alive.Alive, cycle uint) (crd.Crd, bool) {
 	if parent := self.GetParent(); parent != nil {
 		a.direction.SetCrd(parent.GetCrd())
 	}
